services/user-service/requirements.txt
fastapi
uvicorn[standard]
motor
bcrypt
python-jose[cryptography]
pydantic[email]
python-dotenv